Fixed guest registration, refactored ModalSignIn, ModaSignUp, GuestSignUpForm. Fixed comments
