+++ src/client/locales/default.json
@@ -1012,5 +1012,6 @@
   "terms_and_policy_agreement": "I have read and agree to the {Terms}, the {Privacy}, the {Cookies}.",
   "nickname": "Nickname",
   "preferred_languages": "Preferred languages",
-  "public_profile": "Public profile"
+  "public_profile": "Public profile",
+  "need_confirm_agreement": "You need to confirm agreement"
 }
