+++ src/client/locales/ru-RU.json
@@ -1014,5 +1014,6 @@
   "terms_and_policy_agreement": "Я прочитал и согласен с {Terms}, {Privacy} и {Cookies}.",
   "nickname": "Никнейм",
   "preferred_languages": "Предпочитаемые языки",
-  "public_profile": "Публичный профиль"
+  "public_profile": "Публичный профиль",
+  "need_confirm_agreement": "Вам нужно подтвердить согласие"
 }
